refactor: implement basic debug logging
Didn't really have any good way to write log information to aid
debugging before. Did a really basic integration with tracing to get
some developer output. This output is only enabled in debug mode, the
intention is to just `tail -f` the logfile.

This logging setup is not intended for real user information, only for
developers.