feat(config): read repo config for type/scope documentation
commit-lsp now loads the `.commit-lsp.toml` file inside of the
repository root. If this file exists, then it will be used to access
commit type and scope definitions.
Currently this information is only used to provide hover support,
so that summary and full description of the types/scopes are shown on
hover.
In the future this configuration will also be considered for
autocompletion and linting.